requirements.txt
psycopg2-binary==2.9.9
python-dotenv==1.0.0
apscheduler==3.10.4
google-api-python-client==2.108.0
google-auth==2.25.2
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
requests==2.31.0
langchain==0.1.0
langchain-google-genai==0.0.5
langchain-community==0.0.10
langchain-core==0.1.10
langchain-huggingface==0.0.1
faiss-cpu==1.7.4
sentence-transformers==2.2.2
dateparser==1.2.0
